requirements.txt
fastmcp>=0.9.0
pydantic>=2.0.0
httpx>=0.25.0
mcp>=0.9.0
python-dotenv>=1.0.0

